// packages/next/components/DropzoneUploader/types.ts
import {UploadResponse} from '@lib/services/uploadService'

// Types were originally in <DropzoneUploader/> but were moved here do to error messages populating console. Possibly due to introduction of forwardRef use in component.
export interface UploadedFile {
  name: string
  type: string
  lastModified: number
  size: number
  serverResponse: UploadResponse
  originalName?: string
  ext?: string
}

// Marking ext and previewUrl as maybe null helps with prop type checking in <UploadRejectedDialog/>.
export interface DroppedFile {
  name: string
  type: string
  lastModified: number
  size: number
  originalName?: string
  ext: string
  previewUrl: string
}